requirements.txt
torch
torchvision
albumentations
pandas
tqdm
segmentation-models-pytorch==0.3.4
pandas[excel]
tifffile
matplotlib
torchmetrics
pytorch-lightning==1.8.6
